Refactor
- Move the Retry function into the util package.
- Make tar_writer use the common util.Retry function
