+++ pkg/imgpkg/imagetar/tar_file.go
@@ -12,6 +12,7 @@ import (
 
 	regv1 "github.com/google/go-containerregistry/pkg/v1"
 	"github.com/k14s/imgpkg/pkg/imgpkg/imagedesc"
+	"github.com/k14s/imgpkg/pkg/imgpkg/util"
 )
 
 type tarFile struct {
@@ -26,14 +27,6 @@ type tarFileChunk struct {
 }
 
 var _ imagedesc.LayerContents = tarFileChunk{}
-
-type TarEntryNotFoundError struct {
-	Message string
-}
-
-func (t TarEntryNotFoundError) Error() string {
-	return t.Message
-}
 
 type tarFileChunkReadCloser struct {
 	DebugID string
@@ -77,7 +70,7 @@ func (f tarFile) openChunk(path string) (io.ReadCloser, error) {
 				Reader:  tf, Closer: file}, nil
 		}
 	}
-	return nil, TarEntryNotFoundError{fmt.Sprintf("file %s not found in tar. hint: This may be because when copying to a tarball, the --include-non-distributable flag should have been provided.", path)}
+	return nil, util.NonRetryableError{Message: fmt.Sprintf("file %s not found in tar. hint: This may be because when copying to a tarball, the --include-non-distributable flag should have been provided.", path)}
 }
 
 func (f tarFileChunkReadCloser) Close() error {
